issue 1

+++ admin/include/db-favourite.php
@@ -10,7 +10,7 @@ class Favourite extends Table {
 		
 		global $db;
 
-		$filter = "h.active=1 AND h.id IN (SELECT f.ad_id FROM ".static::$table_name." f WHERE f.user_id='$user_id')";
+		$filter = sprintf( "h.active=1 AND h.id IN (SELECT f.ad_id FROM ".static::$table_name." f WHERE f.user_id=%d)", $user_id );
 				
 		return Ad::get_all( array(), $filter, $limit);
 		
@@ -20,7 +20,7 @@ class Favourite extends Table {
 		
 		global $db;
 		
-		$filter = "active=1 AND id IN (SELECT f.ad_id FROM ".static::$table_name." f WHERE f.user_id='$user_id')";
+		$filter = sprintf( "active=1 AND id IN (SELECT f.ad_id FROM ".static::$table_name." f WHERE f.user_id=%d)", $user_id );
 		
 		return Ad::count( array(), $filter );
 		
@@ -30,30 +30,29 @@ class Favourite extends Table {
 		
 		global $db;
 
-		$db->sql = "INSERT INTO ".static::$table_name." (id,user_id,ad_id) 
-		            VALUES ('0','$user_id','$ad_id')";				
-		$db->query();	
+		$db->sql = sprintf( "INSERT INTO ".static::$table_name." (id,user_id,ad_id) VALUES ('0',%d,%d)", $user_id, $ad_id );	
+		$db->query();
 	}
 
 	static public function exists( $user_id, $ad_id ) {
 
 		global $db;
-
-		$db->sql = "SELECT id 
-					FROM ".static::$table_name." 
-					WHERE user_id='$user_id' AND ad_id='$ad_id'
-					LIMIT 1";
-		$db->query();
 		
-		return mysql_num_rows( $db->rs );	
+		$db->sql = sprintf( "SELECT id 
+							 FROM ".static::$table_name." 
+							 WHERE user_id=%d AND ad_id=%d
+							 LIMIT 1",
+							 $user_id, $ad_id );
+		$db->query();
+                 		
+		return mysqli_num_rows( $db->rs );
 	}
 		
 	static public function delete ( $user_id, $ad_id ) {
 		
 		global $db;
 
-		$db->sql = "DELETE FROM ".static::$table_name." 
-		            WHERE user_id='$user_id' AND ad_id='$ad_id'"; 
+		$db->sql = sprintf( "DELETE FROM ".static::$table_name." WHERE user_id=%d AND ad_id=%d", $user_id, $ad_id ); 
 		$db->query();		
 	}
